Coral-AciDevMCP-Agent/pyproject.toml
[project]
name = "firecrawl-coral-agent"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "pydantic-ai[logfire]>=0.4.2",
    "pydantic-ai-slim[mcp]>=0.4.2",
    "Pillow>=10.0.0",
    "python-dotenv>=1.0.0",

]
